Fixes parallel performance for MULE-743, no longer uses "synchronized".
This will fix the situation where transformation is I/O
bound or where the system has multiple CPUs.

You can tune the number of allowable idle Transformer objects by
setting the maxIdleTransformers property on the transformer.
The default is 2 and it won't go lower than 1.

It still constructs a Transformer on initialisation, to check
the source.

git-svn-id: https://svn.codehaus.org/mule/trunk@2034 bf997673-6b11-0410-b953-e057580c5b09