fixed MULE-1093
the issue was that when subsequent calls were being made to axis for example, the event at the 2nd call will contain properties from the first call. Some of these properties will not affect the call, but others like Content-Length.... will eventually mess everything up when the custom properties are added to the soap call.

git-svn-id: https://svn.codehaus.org/mule/trunk/mule@3594 bf997673-6b11-0410-b953-e057580c5b09