MULE-3549: loan broker with cxf fails intermittently
This was a problem with the HTTP transport. When using HTTP 1.1,
servers must send the "Connection: close" header when closing the
connection as keep alives are assumed enabled by default.

Because we weren't doing this, the commons HTTP client was writing
a second message to the existing HTTP connection, when the server
had stopped reading data from it and was in the process of closing
the connection.

I added a DispatchTestCase which was able to reproduce this reliably.




git-svn-id: https://svn.codehaus.org/mule/branches/mule-2.x@12656 bf997673-6b11-0410-b953-e057580c5b09