Jmx changes
Service statistics are now a strict superset of flow statistics.  All flows have a group of "processing" statistics, which represent elapsed time spent in branches of the flow.  Services have, in addition, a group of "execution" statistics, which represent elapsed time spent in the service's component.  The "minimum processing time" statistic is currently always zero, and may be removed in a future integration.


git-svn-id: https://svn.codehaus.org/mule/branches/mule-3.x@20288 bf997673-6b11-0410-b953-e057580c5b09