Refactor security filters so we can have both AuthenticationFilters and filters which perform authorization. To do this I refactored EndpointSecurityFilter into AuthenticationFilter and a generic SecurityFilter. SecurityMessageProcessor will now work with all of these.
In addition to this, I implemented an AuthorizationFilter for Spring Security (EE-1885). It uses the GrantedAuthoritys from the Spring Security information to perform authorization. When a function isn't permitted, you can throw the NotPermittedException which will get turned into an HTTP 405 status code.

This should all be safe to merge into 3.1.x I hope. I saved the other changes for a separate commit.


git-svn-id: https://svn.codehaus.org/mule/branches/mule-3.x@20866 bf997673-6b11-0410-b953-e057580c5b09