Added flow configurations; Changed test cases to load the 2 configs, the flow and the service config.
AbstractServletTestCase (in jersey) -> changed it's parent to AbstractServiceAndFlowTestCase to be able to load flow and service configs in children.

NonRootServletTestCase (in cxf) -> added constructor (because parent's extends was changed to AbstractServiceAndFlowTestCase).

NonRootServletTestCase (in jersey) -> added constructor (because parent's extends was changed to AbstractServiceAndFlowTestCase); Added @Parameters annotated method returning the service and flow configs for the test.



git-svn-id: https://svn.codehaus.org/mule/branches/mule-3.x@22477 bf997673-6b11-0410-b953-e057580c5b09